refactor: modularize gallery component with proper file structure and style organization
