+++ src/components/gallery/MyImageGallery.scss
@@ -1,8 +1,8 @@
-@use '../../styles/variables' as vars;
-@use '../../styles/container' as container;
+// src/components/gallery/MyImageGallery.scss
+@use '../../styles/_variables.scss' as vars;
+@use '../../styles/_container.scss' as container;
 @use "sass:color";
 
-// Animation keyframes
 @keyframes fadeSlideUp {
   0% {
     opacity: 0;
@@ -36,148 +36,5 @@
     @media (max-width: vars.$breakpoint-sm) {
       font-size: 2.5rem;
     }
-  }
-
-  &-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-    grid-gap: 16px;
-    opacity: 0;
-    transform: translateY(30px);
-    animation: fadeSlideUp 0.6s ease forwards;
-    animation-delay: 0.4s;
-
-    @media (min-width: 750px) {
-      grid-template-columns: repeat(2, 1fr);
-    }
-
-    @media (min-width: 900px) {
-      grid-template-columns: repeat(3, 1fr);
-    }
-  }
-
-  &-item {
-    margin-bottom: 1rem;
-    border-radius: 0.5rem;
-    overflow: hidden;
-    background-color: color.adjust(vars.$color-background, $lightness: -2%);
-    box-shadow: 0 4px 6px vars.$box-shadow-color;
-    transition: transform 0.3s ease, box-shadow 0.3s ease;
-    opacity: 0;
-    transform: translateY(20px);
-    transition: all 0.6s cubic-bezier(0.4, 0, 0.2, 1);
-
-    &:hover {
-      transform: translateY(-5px);
-      box-shadow: 0 8px 12px vars.$box-shadow-color;
-    }
-
-    &.visible {
-      opacity: 1;
-      transform: translateY(0);
-    }
-
-    @for $i from 1 through 12 {
-      &:nth-child(#{$i}) {
-        transition-delay: #{$i * 0.1}s;
-      }
-    }
-
-    .gallery-image {
-      width: 100%;
-      height: 200px;
-      object-fit: cover;
-      display: block;
-      opacity: 0;
-      transition: opacity 0.3s ease;
-
-      &.loaded {
-        opacity: 1;
-      }
-    }
-
-    .placeholder {
-      filter: blur(5px);
-      position: absolute;
-      top: 0;
-      left: 0;
-      z-index: 1;
-    }
-  }
-}
-
-.image-container {
-  position: relative;
-  width: 100%;
-  background-color: color.adjust(vars.$color-background, $lightness: -5%);
-  overflow: hidden;
-
-  &.loaded {
-    background-color: transparent;
-  }
-}
-
-.video-container {
-  position: relative;
-  width: 100%;
-  padding-bottom: 56.25%;
-  background-color: color.adjust(vars.$color-background, $lightness: -10%);
-
-  .gallery-video {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    border: none;
-  }
-}
-
-.error-state {
-  padding: 2rem;
-  text-align: center;
-  background-color: color.adjust(vars.$color-accent, $lightness: 35%);
-  color: vars.$color-accent;
-  font-family: vars.$font-secondary;
-  font-size: 0.875rem;
-
-  button {
-    margin-top: 1rem;
-    padding: 0.5rem 1rem;
-    border: none;
-    background-color: vars.$color-primary;
-    color: #fff;
-    font-family: vars.$font-secondary;
-    border-radius: 0.25rem;
-    cursor: pointer;
-    transition: background-color 0.3s;
-
-    &:hover {
-      background-color: color.scale(vars.$color-primary, $lightness: -10%);
-    }
-  }
-}
-
-@keyframes pulse {
-  0%, 100% {
-    opacity: 0.6;
-  }
-  50% {
-    opacity: 1;
-  }
-}
-
-.loading-content {
-  animation: pulse 2s infinite;
-}
-
-@media (prefers-reduced-motion: reduce) {
-  .gallery-item,
-  .gallery-title,
-  .gallery-grid {
-    transition: none;
-    animation: none;
-    opacity: 1;
-    transform: none;
   }
 }
