+++ src/components/gallery/components/ErrorState.scss
@@ -1,0 +1,28 @@
+// src/components/gallery/components/ErrorState.scss
+@use '../../../styles/_variables.scss' as vars;
+@use "sass:color";
+
+.error-state {
+  padding: 2rem;
+  text-align: center;
+  background-color: color.adjust(vars.$color-accent, $lightness: 35%);
+  color: vars.$color-accent;
+  font-family: vars.$font-secondary;
+  font-size: 0.875rem;
+
+  button {
+    margin-top: 1rem;
+    padding: 0.5rem 1rem;
+    border: none;
+    background-color: vars.$color-primary;
+    color: #fff;
+    font-family: vars.$font-secondary;
+    border-radius: 0.25rem;
+    cursor: pointer;
+    transition: background-color 0.3s;
+
+    &:hover {
+      background-color: color.scale(vars.$color-primary, $lightness: -10%);
+    }
+  }
+}
