+++ src/components/gallery/components/GalleryGrid.scss
@@ -1,0 +1,43 @@
+// src/components/gallery/components/GalleryGrid.scss
+@use '../../../styles/_variables.scss' as vars;
+@use "sass:color";
+
+.gallery-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+  grid-gap: 16px;
+  opacity: 0;
+  transform: translateY(30px);
+  animation: fadeSlideUp 0.6s ease forwards;
+  animation-delay: 0.4s;
+
+  @media (min-width: 750px) {
+    grid-template-columns: repeat(2, 1fr);
+  }
+
+  @media (min-width: 900px) {
+    grid-template-columns: repeat(3, 1fr);
+  }
+}
+
+.gallery-item {
+  margin-bottom: 1rem;
+  border-radius: 0.5rem;
+  overflow: hidden;
+  background-color: color.adjust(vars.$color-background, $lightness: -2%);
+  box-shadow: 0 4px 6px vars.$box-shadow-color;
+  transition: transform 0.3s ease, box-shadow 0.3s ease;
+  opacity: 0;
+  transform: translateY(20px);
+  transition: all 0.6s cubic-bezier(0.4, 0, 0.2, 1);
+
+  &:hover {
+    transform: translateY(-5px);
+    box-shadow: 0 8px 12px vars.$box-shadow-color;
+  }
+
+  &.visible {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
