+++ src/components/gallery/components/LoadingState.jsx
@@ -1,0 +1,11 @@
+// src/components/gallery/components/LoadingState.jsx
+import './LoadingState.scss';
+
+export const LoadingState = () => (
+  <div className="gallery-loading">
+    <div className="loading-content">
+      <span className="loading-text">Preparing your visual journey...</span>
+      <div className="loading-progress">Loading amazing moments from paradise</div>
+    </div>
+  </div>
+);
